Refined boss
CHANGELOG:
Added in a new type of drone for Gabi that shoots missiles.
Made Gabi more spammy.
Added in a slight time-freeze effect when you successfully hit an enemy.

TODO:
Rework the melee attack to only destroy bullets with a charge attack
with a very noticavble delay. As of right now, it's too easy to cheese
the boss by destroying every bullet as soon as they fire them.