Code optimization
CHANGELOG:
Made dashing more fluid and responsive
Also tried to optimize Shooting and locking on code. Resulted in bugs.

NEW BUGS:
When shooting a weapon with a noticable start up delay, the player does
not STAY facing in the angle towards a locked on enemy.
Fix PlayerShooting; allow guns with the same control be allowed to fire
with each other, while disallowing another gun. (Possible fix: check in
PlayerManager what input button is pressed and only allow for barrels
with the correct input to fire)